backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
defusedxml==0.7.1
Django==4.2.7
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
djoser==2.2.0
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==7.3.0
gunicorn==21.2.0
idna==3.11
mccabe==0.7.0
oauthlib==3.3.1
packaging==26.1
Pillow==10.1.0
psycopg2-binary==2.9.9
pycodestyle==2.14.0
pycparser==3.0
pyflakes==3.4.0
PyJWT==2.12.1
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2026.1.post1
requests==2.33.1
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.8.5
sqlparse==0.5.5
tzdata==2026.1
urllib3==2.6.3
